+++ src/Domains/Inventory/Variants/Models/VariantsWarehousesPriceHistory.php
@@ -35,7 +35,7 @@ class VariantsWarehousesPriceHistory extends BaseModel
         ];
     }
 
-    public function getPriceAttribute(string|int $value): float
+    public function getPriceAttribute(string|float $value): float
     {
         return (float) $value;
     }
